committing rhyme api hint, where it takes the target selected word and generates a rhyming word as hint for user. Current issue where it seems out of date due to multiple api calls, need to debug
