+++ src/App.js
@@ -1,13 +1,13 @@
 import logo from './logo.svg';
 import React from 'react';
 import './App.css';
-import { fetchData } from './services/quotes-fetch';
+import { fetchData } from './services/quotesFetch';
 import { useState, useEffect } from 'react';
-import { getQuotes } from './services/quotes-fetch'
+import { getQuotes } from './services/quotesFetch'
 import { getRoles } from '@testing-library/react';
-import DisplayQuote from './tests/Quote-Display/displayquote';
-import DisplayRhyme from './tests/Rhyme-Display/displayrhyme';
-import DisplaySynonym from './tests/Thesaurus-Display/displaythesaurus';
+import DisplayQuote from './tests/Quote-Display/displayQuote';
+import DisplayRhyme from './tests/Rhyme-Display/displayRhyme';
+import DisplaySynonym from './tests/Thesaurus-Display/displayThesaurus';
 
 
 function App() {
